pipelines/pytests/requirements.txt
pytest
azure-identity
azure-mgmt-resource



pandas
pyarrow
pytest-azurepipelines
aiohttp
pyodbc
